移动 Cloud.define 等函数到 cloud.js; 移动 Cloud.httpRequest 到 http-request.js
