+++ lib/leanengine.js
@@ -19,6 +19,7 @@ if (process.env.LC_API_SERVER) {
 var NODE_ENV = process.env.NODE_ENV || 'development';
 AV.Cloud.__prod = NODE_ENV === 'production' ? 1 : 0;
 
+// override AV.Cloud to a connect app
 if (!AV._old_Cloud) {
   AV._old_Cloud = AV.Cloud;
   var Cloud = AV.Cloud = connect();
@@ -26,6 +27,8 @@ if (!AV._old_Cloud) {
     Cloud[key] = AV._old_Cloud[key];
   }
 }
+
+AV._.extend(AV.Cloud, require('./cloud'));
 
 // Don't reject unauthorized ssl.
 if (https.globalAgent && https.globalAgent.options) {
@@ -297,24 +300,6 @@ var respError = function(res, err) {
   }));
 };
 
-Cloud.run = function(name, data, options) {
-  var promise = new AV.Promise();
-  try {
-    Cloud.__code[name]({params: data, user: AV.User.current()}, {
-      success: function(result) {
-        promise.resolve(result);
-      },
-      error: function(err) {
-        promise.reject(err);
-      }
-    });
-  } catch (err) {
-    console.log('Run function \'' + name + '\' failed with error:', err);
-    promise.reject(err);
-  }
-  return promise._thenRunCallbacks(options);
-};
-
 var call = function(funcName, params, user, meta, options, cb) {
   if (!options) {
     options = {};
@@ -455,13 +440,6 @@ var onCompleteBigQueryJob = function(data) {
   }
 };
 
-Cloud.__code = {};
-
-Cloud.define = function(name, func) {
-  debug('define function: %s', name);
-  Cloud.__code[name] = func;
-};
-
 var hookNameMapping = {
   beforeSave: '__before_save_for_',
   beforeUpdate: '__before_update_for_',
@@ -469,80 +447,6 @@ var hookNameMapping = {
   afterUpdate: '__after_update_for_',
   beforeDelete: '__before_delete_for_',
   afterDelete: '__after_delete_for_'
-};
-
-var _define = function(className, hook, func) {
-  debug('define class hook: %s %s', hook, className);
-  Cloud.__code[hook + className] = func;
-};
-
-Cloud.beforeSave = function(nameOrClass, func) {
-  _define(className(nameOrClass), '__before_save_for_', func);
-};
-
-Cloud.afterSave = function(nameOrClass, func) {
-  _define(className(nameOrClass), '__after_save_for_', func);
-};
-
-Cloud.beforeUpdate = function(nameOrClass, func) {
-  _define(className(nameOrClass), '__before_update_for_', func);
-};
-
-Cloud.afterUpdate = function(nameOrClass, func) {
-  _define(className(nameOrClass), '__after_update_for_', func);
-};
-
-Cloud.beforeDelete = function(nameOrClass, func) {
-  _define(className(nameOrClass), '__before_delete_for_', func);
-};
-
-Cloud.afterDelete = function(nameOrClass, func) {
-  _define(className(nameOrClass), '__after_delete_for_', func);
-};
-
-Cloud.onVerified = function(type, func) {
-  Cloud.define('__on_verified_' + type, func);
-};
-
-Cloud.onLogin = function(func) {
-  Cloud.define('__on_login__User', func);
-};
-
-
-if (!AV.Insight) {
-  AV.Insight = {};
-}
-
-AV.Insight.on = function(action, func) {
-  Cloud.define('__on_complete_bigquery_job', func);
-};
-
-Cloud.logInByIdAndSessionToken = function(uid, sessionToken, fetch, cb) {
-  if (fetch) {
-    AV.User.become(sessionToken, {
-      success: function(user) {
-        return cb(null, user);
-      },
-      error: function(user, err) {
-        return cb(err);
-      }
-    });
-  } else {
-    var user = new AV.User();
-    user.id = uid;
-    user._sessionToken = sessionToken;
-    AV.User._saveCurrentUser(user);
-    return cb(null, user);
-  }
-};
-
-var className = function(clazz) {
-  if (utils.typeOf(clazz) === 'string') {
-    return clazz;
-  }
-  if (clazz.className)
-    return clazz.className;
-  throw new Error("Unknown class:" + clazz);
 };
 
 var unauthResp = function(res) {
